+++ NewPlatform.Flexberry.ORM.ODataService/Handlers/PostPatchHandler.cs
@@ -1,6 +1,10 @@
 ﻿#if NETFRAMEWORK
 namespace NewPlatform.Flexberry.ORM.ODataService.Handlers
 {
+    using Microsoft.AspNet.OData.Batch;
+    using System;
+    using System.IO;
+    using System.Net;
     using System.Net.Http;
     using System.Threading;
     using System.Threading.Tasks;
@@ -33,8 +37,13 @@ namespace NewPlatform.Flexberry.ORM.ODataService.Handlers
         public const string AcceptApplicationMsExcel = "PostPatchHandler_AcceptApplicationMsExcel";
 
         /// <inheritdoc/>
-        protected override Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
+        protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
         {
+            if (request == null)
+            {
+                throw new ArgumentNullException(nameof(request));
+            }
+
             foreach (var val in request.Headers.Accept)
             {
                 if (val.MediaType == "application/ms-excel")
@@ -64,7 +73,52 @@ namespace NewPlatform.Flexberry.ORM.ODataService.Handlers
             if (!request.Headers.Contains("Accept-Charset"))
                 request.Headers.Add("Accept-Charset", new[] { "utf-8" });
 
-            return base.SendAsync(request, cancellationToken);
+            HttpResponseMessage responseMessage = await base.SendAsync(request, cancellationToken);
+            await RewriteResponse(responseMessage);
+            return responseMessage;
+
+        }
+
+        /// <summary>
+        /// Удаление некорректных символов в именах типов метаданных.
+        /// Данные символы добавляются специально, чтобы MS не бросало исключение на пустой Namespace у типов с PublishName.
+        /// </summary>
+        /// <param name="response">Текущее ответное сообщение, в котором могут быть заменены символы.</param>
+        /// <returns>Задача на обработку.</returns>
+        private static async Task RewriteResponse(HttpResponseMessage response)
+        {
+            /* Same code for NETSTANDARD is placed on NewPlatform.Flexberry.ORM.ODataService.Extensions.ODataApplicationBuilderExtensions.*/
+            string contentType = response?.Content?.Headers?.ContentType?.MediaType;
+            if (!string.IsNullOrEmpty(contentType) &&
+            (contentType.Contains("application/json")
+            || contentType.Contains("application/xml")
+            || contentType.Contains("multipart/mixed")))
+            {
+                HttpContent content = response.Content;
+                Stream contentStream = await content.ReadAsStreamAsync();
+                string responseStr = new StreamReader(contentStream).ReadToEnd();
+                if (!string.IsNullOrEmpty(responseStr) && responseStr.Length > 3)
+                {
+                    responseStr = responseStr
+                   .Replace("(____.", "(")
+                   .Replace("\"____.", "\"")
+                   .Replace("____.", ".")
+                   .Replace(" Namespace=\"____\"", " Namespace=\"\"");
+
+                    using (MemoryStream newStream = new MemoryStream())
+                    {
+                        using StreamWriter sw = new StreamWriter(newStream);
+                        sw.Write(responseStr);
+                        sw.Flush();
+                        newStream.Seek(0, SeekOrigin.Begin);
+
+                        contentStream.Position = 0;
+                        contentStream.SetLength(responseStr.Length);
+
+                        await newStream.CopyToAsync(contentStream);
+                    }
+                }
+            }
         }
     }
 }